import optimisation for moved files
